requirements.txt
streamlit~=1.41.1
pandas~=2.2.3
pytz~=2024.2
plotly~=5.24.1
matplotlib~=3.10.0
reportlab~=4.2.5
pillow~=11.1.0
bcrypt~=4.2.1
tensorflow~=2.18.0
numpy~=2.0.2
tldextract~=5.1.3
validators~=0.34.0
scikit-learn~=1.6.1
imblearn~=0.0
imbalanced-learn~=0.13.0
keras~=3.8.0
requests~=2.32.3
python-owasp-zap-v2.4